scraper_backend/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
playwright>=1.40.0
supabase>=2.3.0
pydantic>=2.5.2
python-dotenv>=1.0.0
